refactor: Transfer TestController inner class to class
